Updated HUD, added some messaging.
FEATURE: Added messaging that appears upon the start of the game to give some directions.
FEATURE: Added messaging that appears when users collect all the coins.
FIX: Moved elements in the HUD (key icon, coin and score) to be in a better position that was less likely to strain eyes.